- Refactor types - Expose types in build output
