+++ package.json
@@ -7,6 +7,7 @@
   "private": false,
   "repository": "nightborn-be/signum",
   "main": "dist/index.js",
+  "types": "./dist/index.d.ts",
   "module": "dist/index.es.js",
   "jsnext:main": "dist/index.es.js",
   "files": [
