--- src/components/chat-body/interfaces/IChatBodyInformation.ts
@@ -1,8 +1,0 @@
-import { IOption } from "../../../interfaces/IOption";
-
-export interface IChatBodyInformation {
-    email: string,
-    message: string,
-    steps: IOption[],
-    option: IOption
-}
