--- src/components/chat-context/interfaces/IChatProviderProps.ts
@@ -1,8 +1,0 @@
-import { IOption } from "../../../interfaces/IOption";
-import { IChatConfig } from "../../../interfaces/IChatConfig";
-
-export interface IChatProviderProps {
-    option: IOption,
-    config: IChatConfig,
-    children: any
-}
