+++ src/index.tsx
@@ -1,8 +1,15 @@
 import React from 'react';
 import ChatBody from './components/chat-body';
 import ChatIcon from './components/chat-icon';
-import { IChatProps } from './interfaces/IChatProps';
 import ChatContextProvider from './components/chat-context';
+import { IOption, IChatConfig } from './types';
+
+interface IChatProps {
+	option: IOption,
+    config: IChatConfig
+}
+
+export * from './types';
 
 export default function Chat(props: IChatProps) {
 	return (
