--- src/interfaces/IChatProps.ts
@@ -1,7 +1,0 @@
-import { IOption } from "./IOption";
-import { IChatConfig } from "./IChatConfig";
-
-export interface IChatProps {
-    option: IOption,
-    config: IChatConfig
-}
