--- src/interfaces/IOption.ts
@@ -1,8 +1,0 @@
-export interface IOption {
-    title: string,
-    subTitle?: string,
-    message: string,
-    name: string,
-    handleOnClick?: Function,
-    options?: IOption[]
-}
